Adds OVPN file configuration form
Adds a new page for configuring OVPN files.
This allows users to customize the OpenVPN configuration.
Updates the server details page to include a link to the new configuration page.
Increases the default height of the config template textarea
Updates route path of the ovpn config page to the new standrad servers/{id}/ovpn-file-config